-got rid of some Migrations because I ended up performing the actions in the database directly -removed controller tests for frontend because controllers should be skinny. Need to fix that for one method. -implemented SSE to send response from booking microservice to the frontend instead of SignalR -updated ocelot file to map the status update endpoints correctly -added SSE service to bookingMicroservice to handle SSE communication -removed sessionID from models in favor of using appointmentId, which I generate when the appointment request is sent in the frontend. -changes appointmentID to long type instead of int to ensure a unique ID even if generated simultaneously with another ID. -enabled CORS from the bookingMS to get SSE to work - will probably restrict in the future. -added a new column to the Appointment table so that I do not have to recreate the table. The ID field is already an int and has a relationship to patient so I would rather just create a new field AppointmentUniqueId and use it to correctly send the SSE response (routed via the uniqueId) -removed status update endpoint from BookingMS controller since it is sent directly from the SSE service -removed all SignalR references and code -added JS event listener to AppointmentPending page to route the user to the appropriate status page when their appointment's status is sent. -added controller methods to route to the correct status pages (completed or failed)

+++ DotNetProject8/Models/AppointmentResponse.cs
@@ -2,7 +2,7 @@
 {
     public class AppointmentResponse
     {
-        public int AppointmentId { get; set; }
+        public long AppointmentId { get; set; }
         public DateTime AppointmentDate { get; set; } = new DateTime();
         public DateTime AppointmentTime { get; set; } = new DateTime();
         public int ConsultantId { get; set; } = new int();
